refactor: toast example file names correction #305
