+++ apps/doc/src/app/components/toast/examples/message-info-example/message-info-example.component.ts
@@ -1,10 +1,10 @@
 import { Component, TemplateRef } from '@angular/core';
-import { PrizmToastService, PrizmToastAppearance, PrizmToastPosition } from '@prizm-ui/components';
+import { PrizmToastAppearance, PrizmToastPosition, PrizmToastService } from '@prizm-ui/components';
 import { UntypedFormControl } from '@angular/forms';
 
 @Component({
-  selector: 'prizm-message-danger-example',
-  templateUrl: './message-success-example.component.html',
+  selector: 'prizm-message-info-example',
+  templateUrl: './message-info-example.component.html',
   styles: [
     `
       .box {
@@ -20,6 +20,11 @@ import { UntypedFormControl } from '@angular/forms';
 
       .footer {
         margin-top: 8px;
+      }
+
+      .inline-container {
+        border: 1px solid black;
+        padding: 10px;
       }
 
       .date {
@@ -41,7 +46,9 @@ import { UntypedFormControl } from '@angular/forms';
     `,
   ],
 })
-export class PrizmToastDangerExampleComponent {
+export class PrizmToastInfoExampleComponent {
+  readonly containerId = 'inline-container-info';
+
   readonly data = [
     {
       val: PrizmToastPosition.TOP_MIDDLE,
@@ -67,9 +74,13 @@ export class PrizmToastDangerExampleComponent {
       val: PrizmToastPosition.BOTTOM_RIGHT,
       label: 'Bottom Right',
     },
+    {
+      val: this.containerId,
+      label: 'В строковом контейнере',
+    },
   ];
 
-  appearance: PrizmToastAppearance = 'danger';
+  appearance: PrizmToastAppearance = 'info';
   readonly formControl = new UntypedFormControl(PrizmToastPosition.TOP_RIGHT);
 
   constructor(private readonly toastService: PrizmToastService) {}
@@ -78,60 +89,85 @@ export class PrizmToastDangerExampleComponent {
     this.toastService.create('Hello #' + Math.random(), {
       appearance: this.appearance,
       position: this.formControl.value,
-      timer: 5000,
       title: 'Заголовок',
     });
   }
 
   public showToastWithoutTitle(): void {
     this.toastService.create('Старайтесь уместить текст в 1 строку.', {
+      timer: 5000,
       appearance: this.appearance,
       position: this.formControl.value,
-      timer: 5000,
     });
   }
 
   public showToastWithoutTimer(): void {
     this.toastService.create('Это сообщение будет показываться пока не закроете.', {
-      appearance: this.appearance,
-      position: this.formControl.value,
       timer: 0,
+      appearance: this.appearance,
       title: 'Заголовок',
+      position: this.formControl.value,
     });
+  }
+
+  public showWithId(): void {
+    this.toastService.create('Это сообщение будет показываться пока не закроете.', {
+      timer: 0,
+      id: 'our-id-1',
+      appearance: this.appearance,
+      title: 'Заголовок',
+      position: this.formControl.value,
+    });
+  }
+
+  public updateContentWithId(): void {
+    this.toastService.updateContent('our-id-1', 'Обновили содержимое №' + Math.random());
+  }
+
+  public updateTitleWithId(template: TemplateRef<unknown>): void {
+    this.toastService.updateTitle('our-id-1', template);
+  }
+
+  public closeWithId(): void {
+    this.toastService.delete('our-id-1');
+  }
+
+  public closeAll(): void {
+    this.toastService.deleteAll();
   }
 
   public showToastWithBigText(): void {
     this.toastService.create('Lorem Ipsum is simply dummy text of the printing and typesetting industry.', {
-      appearance: this.appearance,
-      position: this.formControl.value,
       timer: 5000,
+      appearance: this.appearance,
       title: 'Большой заголовок очень очень очень очень',
+      position: this.formControl.value,
     });
   }
 
   public showToastWithContentTemplate(contentTemplate: TemplateRef<unknown>): void {
     this.toastService.create(contentTemplate, {
-      appearance: this.appearance,
-      position: this.formControl.value,
       timer: 5000,
+      appearance: this.appearance,
       title: 'Большой заголовок очень очень очень очень',
+      position: this.formControl.value,
     });
   }
 
   public showToastWithHeaderTemplate(headerTemplate: TemplateRef<unknown>): void {
     this.toastService.create('Шаблон в хедере', {
+      timer: 0,
       appearance: this.appearance,
-      position: this.formControl.value,
-      timer: 5000,
       title: headerTemplate,
+      position: this.formControl.value,
     });
   }
 
   public showToastWithBigTitle(): void {
     this.toastService.create('Содержимое', {
+      timer: 5000,
       appearance: this.appearance,
       position: this.formControl.value,
-      timer: 5000,
       title:
         " Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
     });
